Wait a certain amount of time for ART to process
A stopgap solution to prevent the commonly occurring problem where the
Proctor is unable to login.

The Adapter needs to wait for ART to processes the pre-loaded data
and calculate proctor assessments. Proctor calls ART for assessments
then stores the results into the database for the Proctor->GetInitData 
invocation made soon after logging in.

JIRA: CP-403