Add an EXTRACTION phase to the automation
After simulating a test session and before cleaning up, this commit
will extract TDS Reports. This is necessary because TIS will query
ART for student accommodation data. If cleanup was performed before
extracting TDS Reports, then TIS would fail in that regard.

The TDS Extraction code was refactored out of the 
SbossAutomationAdapterService class into its own class that can be
called from both the service class and the AutomationTaskRunner.

Extraction is called twice. Once immediately after the test simulation
phase and again once automation is complete. The second call the 
extraction should return quickly since the data already exists in the
TIS database and the client won’t need to wait again for TIS to compute
final TDS Reports. Also, it’s the data from the second call to the 
extraction method that is ultimately used by the client.